+++ lib/declare_schema/model/index_definition.rb
@@ -19,7 +19,7 @@ module DeclareSchema
         @table = options.delete(:table_name) || model.table_name
         @fields = Array.wrap(fields).map(&:to_s)
         @explicit_name = options[:name] unless options.delete(:allow_equivalent)
-        @name = options.delete(:name) || model.connection.index_name(table, column: @fields).gsub(/index.*_on_/, 'on_')
+        @name = options.delete(:name) || self.class.index_name(@fields)
         @unique = options.delete(:unique) || name == PRIMARY_KEY_NAME || false
 
         if @name.length > MYSQL_INDEX_NAME_MAX_LENGTH
@@ -58,6 +58,10 @@ module DeclareSchema
             index_definitions << new(model, primary_key_columns, name: PRIMARY_KEY_NAME, unique: true, where: nil, table_name: old_table_name)
           end
           index_definitions
+        end
+
+        def index_name(columns)
+          "on_#{Array(columns).join("_and_")}"
         end
 
         private