+++ spec/lib/declare_schema/migration_generator_spec.rb
@@ -373,14 +373,14 @@ RSpec.describe 'DeclareSchema Migration Generator' do
 
         add_index :adverts, [:category_id], name: 'on_category_id'
 
-        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" if defined?(Mysql2)}
+        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" if defined?(Mysql2)}
       EOS
       .and migrate_down(<<~EOS.strip)
         remove_column :adverts, :category_id
 
         remove_index :adverts, name: :on_category_id rescue ActiveRecord::StatementInvalid
 
-        #{"remove_foreign_key(\"adverts\", name: \"index_adverts_on_category_id\")\n" if defined?(Mysql2)}
+        #{"remove_foreign_key(\"adverts\", name: \"on_category_id\")\n" if defined?(Mysql2)}
       EOS
     )
 
@@ -401,8 +401,8 @@ RSpec.describe 'DeclareSchema Migration Generator' do
 
         add_index :adverts, [:c_id], name: 'on_c_id'
 
-        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" +
-          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"index_adverts_on_c_id\")" if defined?(Mysql2)}
+        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" +
+          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"on_c_id\")" if defined?(Mysql2)}
       EOS
     )
 
@@ -421,8 +421,8 @@ RSpec.describe 'DeclareSchema Migration Generator' do
       migrate_up(<<~EOS.strip)
         add_column :adverts, :category_id, :integer, limit: 8, null: false
 
-        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" +
-          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"index_adverts_on_c_id\")" if defined?(Mysql2)}
+        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" +
+          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"on_c_id\")" if defined?(Mysql2)}
       EOS
     )
 
@@ -443,8 +443,8 @@ RSpec.describe 'DeclareSchema Migration Generator' do
 
         add_index :adverts, [:category_id], name: 'my_index'
 
-        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" +
-          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"index_adverts_on_c_id\")" if defined?(Mysql2)}
+        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" +
+          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"on_c_id\")" if defined?(Mysql2)}
       EOS
     )
 
@@ -469,16 +469,16 @@ RSpec.describe 'DeclareSchema Migration Generator' do
         add_column :adverts, :updated_at, :datetime, null: true
         add_column :adverts, :lock_version, :integer#{lock_version_limit}, null: false, default: 1
 
-        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" +
-          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"index_adverts_on_c_id\")" if defined?(Mysql2)}
+        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" +
+          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"on_c_id\")" if defined?(Mysql2)}
       EOS
       .and migrate_down(<<~EOS.strip)
         remove_column :adverts, :created_at
         remove_column :adverts, :updated_at
         remove_column :adverts, :lock_version
 
-        #{"remove_foreign_key(\"adverts\", name: \"index_adverts_on_category_id\")\n" +
-          "remove_foreign_key(\"adverts\", name: \"index_adverts_on_c_id\")" if defined?(Mysql2)}
+        #{"remove_foreign_key(\"adverts\", name: \"on_category_id\")\n" +
+          "remove_foreign_key(\"adverts\", name: \"on_c_id\")" if defined?(Mysql2)}
       EOS
     )
 
@@ -502,8 +502,8 @@ RSpec.describe 'DeclareSchema Migration Generator' do
 
         add_index :adverts, [:title], name: 'on_title'
 
-        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" +
-          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"index_adverts_on_c_id\")" if defined?(Mysql2)}
+        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" +
+          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"on_c_id\")" if defined?(Mysql2)}
       EOS
     )
 
@@ -523,8 +523,8 @@ RSpec.describe 'DeclareSchema Migration Generator' do
 
         add_index :adverts, [:title], unique: true, name: 'on_title'
 
-        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" +
-          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"index_adverts_on_c_id\")" if defined?(Mysql2)}
+        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" +
+          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"on_c_id\")" if defined?(Mysql2)}
       EOS
     )
 
@@ -544,8 +544,8 @@ RSpec.describe 'DeclareSchema Migration Generator' do
 
         add_index :adverts, [:title], name: 'my_index'
 
-        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" +
-          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"index_adverts_on_c_id\")" if defined?(Mysql2)}
+        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" +
+          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"on_c_id\")" if defined?(Mysql2)}
       EOS
     )
 
@@ -563,8 +563,8 @@ RSpec.describe 'DeclareSchema Migration Generator' do
 
         add_index :adverts, [:title], name: 'on_title'
 
-        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" +
-          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"index_adverts_on_c_id\")" if defined?(Mysql2)}
+        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" +
+          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"on_c_id\")" if defined?(Mysql2)}
       EOS
     )
 
@@ -582,8 +582,8 @@ RSpec.describe 'DeclareSchema Migration Generator' do
 
         add_index :adverts, [:title], unique: true, name: 'my_index'
 
-        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" +
-          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"index_adverts_on_c_id\")" if defined?(Mysql2)}
+        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" +
+          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"on_c_id\")" if defined?(Mysql2)}
       EOS
     )
 
@@ -601,8 +601,8 @@ RSpec.describe 'DeclareSchema Migration Generator' do
 
         add_index :adverts, [:title, :category_id], name: 'on_title_and_category_id'
 
-        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" +
-          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"index_adverts_on_c_id\")" if defined?(Mysql2)}
+        #{"add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" +
+          "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"on_c_id\")" if defined?(Mysql2)}
       EOS
     )
 
@@ -638,8 +638,8 @@ RSpec.describe 'DeclareSchema Migration Generator' do
             "add_index :ads, [:id], unique: true, name: 'PRIMARY'\n"
           elsif defined?(Mysql2)
             "execute \"ALTER TABLE ads DROP PRIMARY KEY, ADD PRIMARY KEY (id)\"\n\n" +
-            "add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"index_adverts_on_category_id\")\n" +
-            "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"index_adverts_on_c_id\")"
+            "add_foreign_key(\"adverts\", \"categories\", column: \"category_id\", name: \"on_category_id\")\n" +
+            "add_foreign_key(\"adverts\", \"categories\", column: \"c_id\", name: \"on_c_id\")"
           end}
       EOS
       .and migrate_down(<<~EOS.strip)
@@ -652,8 +652,8 @@ RSpec.describe 'DeclareSchema Migration Generator' do
             "add_index :adverts, [:id], unique: true, name: 'PRIMARY'\n"
           elsif defined?(Mysql2)
             "execute \"ALTER TABLE adverts DROP PRIMARY KEY, ADD PRIMARY KEY (id)\"\n\n" +
-            "remove_foreign_key(\"adverts\", name: \"index_adverts_on_category_id\")\n" +
-            "remove_foreign_key(\"adverts\", name: \"index_adverts_on_c_id\")"
+            "remove_foreign_key(\"adverts\", name: \"on_category_id\")\n" +
+            "remove_foreign_key(\"adverts\", name: \"on_c_id\")"
           end}
       EOS
     )